ctualización de servicio para publicar capas en GeoServer
Instalación de librería file-type para identificación de archivos a
través de su buffer.

Actualización de configuración de app debido a refactorización de 2
variables de ambiente y 2 nuevas variables de ambiente.

Nuevo archivo de configuración de capas de catastro minero.

Nueva migración para cambiar el tipo de columna de geometrías y dejarlas
abiertas o tipo geometry(Geometry,4326).

Refactorización de servicio shaper para publicación de capas.

+++ src/modules/shaper/services/shaper.service.ts
@@ -1,27 +1,20 @@
 import { Injectable, Logger } from '@nestjs/common';
-import path = require('path');
-import os = require('os');
 import { HelperService } from '../../../helper/helper.service';
 import { GeoserverService } from '@services/geoserver.service';
 import { AppConfigService } from 'src/config/config.service';
-import AdmZip from 'adm-zip';
+import * as AdmZip from 'adm-zip';
+import { fromBuffer } from 'file-type';
 
 @Injectable()
 export class ShaperService {
   private G_HOST: string;
-  private TEMDIR: string;
   private readonly logger = new Logger(ShaperService.name);
   constructor(
-    // TODO: eliminado forwardRed 2
     private appConfigService: AppConfigService,
     private geoService: GeoserverService,
     private helperService: HelperService,
   ) {
     this.G_HOST = this.appConfigService.serverHost;
-
-    // get temp directory
-    this.TEMDIR = os.tmpdir(); // /tmp
-    // this.TPLDIR = path.join(__dirname, '..', '..', 'tpl');
 
     if (!this.G_HOST) {
       throw new Error(`GEOSERVER variables are missing`);
@@ -29,50 +22,88 @@ export class ShaperService {
   }
   public async shapeHandler(
     pathAndFile: string,
-    folders3: string,
     nameShapefile: string,
     type: string,
+    style?: string,
   ) {
     this.logger.log(`Iniciando shaperHandler para '${nameShapefile}'`);
 
     const shpLowerCase = nameShapefile.toLowerCase();
     const zipBody = await this.helperService.downloadFileS3(pathAndFile);
 
+    if (!zipBody) {
+      this.logger.error(`Archivo ${nameShapefile}.zip no encontrado en S3`);
+      this.logger.log('Finalizando shaperHandler');
+      return;
+    }
+
     const { capa, inputsGeom } = await this.helperService.getGeometriesAndCapa(
       zipBody,
       shpLowerCase,
     );
 
-    const { inserted, deleted } =
-      await this.helperService.deleteOldAndSaveNewGeometrias(capa, inputsGeom);
-    this.logger.log(
-      `Elimnados ${deleted} filas e insertadas ${inserted} filas`,
+    if (!capa || !inputsGeom) {
+      this.logger.log(`Finalizado sin importar capa: '${nameShapefile}'`);
+      return 'done';
+    }
+
+    await this.helperService.deleteOldAndSaveNewGeometrias(capa, inputsGeom);
+
+    const layername = await this.geoService.getLayerName(
+      shpLowerCase,
+      this.appConfigService.workspaceCatastroMinero,
+      this.appConfigService.datastoreCatastroMinero,
     );
 
-    const layername = await this.geoService.getLayerName(shpLowerCase);
+    if (!layername)
+      await this.geoService.publishLayer(
+        shpLowerCase,
+        type,
+        this.appConfigService.workspaceCatastroMinero,
+        this.appConfigService.datastoreCatastroMinero,
+      );
 
-    if (!layername) await this.geoService.publishLayer(shpLowerCase, type);
+    const styleEntry = await this.geoService.getStyle(`${shpLowerCase}_style`);
 
-    const stylename = await this.geoService.getStyle(`${shpLowerCase}_style`);
-
-    if (!stylename) await this.geoService.createStyle(`${shpLowerCase}_style`);
+    if (!styleEntry) {
+      if (typeof styleEntry === 'undefined')
+        await this.geoService.createStyle(`${shpLowerCase}_style`);
+      this.logger.warn(
+        'Estilo de capa en blanco, actualizando por valor configurado',
+      );
+      await this.geoService.uploadStyle(capa.style, `${shpLowerCase}_style`);
+    }
 
     const zipFile = new AdmZip(zipBody);
-    console.log('1. zipFile:', zipFile);
+    const styleFile = zipFile.getEntry(`${nameShapefile}.sld`);
 
-    const sldfile = path.join(this.TEMDIR, folders3, `${shpLowerCase}.sld`);
-    console.log(sldfile);
-    await this.geoService.uploadStyle(sldfile, `${shpLowerCase}_style`);
+    if (styleFile) {
+      this.logger.log('Archivo de estilos .sld encontrado en .zip');
+      const styleData = styleFile.getData();
+      const styleType = await fromBuffer(styleData);
+
+      if (styleType.mime !== 'application/xml')
+        this.logger.error('Tipo de documento SLD no es XML');
+      else {
+        const stypeString = styleData.toString();
+        await this.geoService.uploadStyle(stypeString, `${shpLowerCase}_style`);
+      }
+    } else {
+      this.logger.warn('Archivo de estilos .sld NO encontrado');
+    }
 
     // Apply Style
-    await this.geoService.setLayerStyle(shpLowerCase, `${shpLowerCase}_style`);
+    const styleName = style ? style : `${shpLowerCase}_style`;
+    await this.geoService.setLayerStyle(
+      shpLowerCase,
+      styleName,
+      this.appConfigService.workspaceCatastroMinero,
+    );
 
-    // const tempdelete = path.join(this.TEMDIR, folders3);
+    this.logger.log(
+      `Finalizado e importado shaperHandler para '${nameShapefile}'.`,
+    );
 
-    //await fs.rmSync(tempdelete, { recursive: true, force: true });
-
-    // Delete layer if need
-    // DELETE http://<url>/geoserver/rest/workspaces/<workspaceName>/coveragestores/<storeName>/coverages/<layerName>?recurse=true
     return 'done';
   }
 }
